requirements.txt
Django>=4.1
psycopg2-binary>=2.9
djangorestframework>=3.14
pandas>=1.5.1
SQLAlchemy>=1.4